adding ids and tests
combined both tests.py and added testingrequirements.txt which has all the selenium base dependencies
