+++ ap_src/templates/Base.html
@@ -139,10 +139,10 @@
       <div class="field is-grouped">
           <div class="buttons">
             {% if user.is_authenticated %}
-                <a href="/profile/" class="button is-info">{{request.user.username}}
+                <a id="profile" href="/profile/" class="button is-info">{{request.user.username}}
                   <i class="fas fa-user-circle ml-2 is-size-5"></i>
                 </a>
-                <a href="{% url 'logout' %}" class="button is-light">Log out</a>
+                <a id="logout" href="{% url 'logout' %}" class="button is-light">Log out</a>
             {% else %}
                 <a id="signup" href="{% url 'signup' %}" class="button is-info">
                   <strong>Sign up</strong>
