+++ ap_src/templates/registration/signup.html
@@ -11,7 +11,7 @@
         
             
         
-        <button type="submit" class="button mt-1 is-link">Sign Up</button>
+        <button id="submit" type="submit" class="button mt-1 is-link">Sign Up</button>
     </form>
 </div>
