+++ ap_src/templates/teams/calendar.html
@@ -38,7 +38,7 @@
                     <a class="button is-primary is-outlined modal-button js-modal-trigger" id="#modal">
                         <i class="fas fa-user-plus"></i>
                     </a>
-                    <a class="button is-info is-outlined mx-1" href="{% url 'team_settings' team.id %}">
+                    <a id="settings" class="button is-info is-outlined mx-1" href="{% url 'team_settings' team.id %}">
                         <i class="fas fa-cog"></i>
                     </a>
                     {% endif %}
