+++ ap_src/templates/teams/create_team.html
@@ -28,7 +28,7 @@
         
 
 
-    <input class="button is-link" type="submit" value="Submit" onclick="makeAlert()">
+    <input id="submit" class="button is-link" type="submit" value="Submit" onclick="makeAlert()">
     </form>
 </div>
