+++ ap_src/templates/teams/settings.html
@@ -45,10 +45,10 @@
                         {% if User.role == member %}
                         <a class="button is-info" href="{% url 'promote_team_member' team.id User.user.id%}">Promote to Co-Owner</a>
                         {% elif User.role == coowner %}
-                        <a class="button is-info" href="{% url 'demote_team_member' team.id User.user.id%}">Demote to Member</a>
+                        <a id="promote_{{User.user.username}}" class="button is-info" href="{% url 'demote_team_member' team.id User.user.id%}">Demote to Member</a>
                         {% endif %}
                         {% if User.role != owner %}
-                        <a class="button is-danger" id="{{User.user.username}}" href="{% url 'remove_team_member' team.id User.user.id %}">Remove</a>
+                        <a class="button is-danger" id="remove_{{User.user.username}}" href="{% url 'remove_team_member' team.id User.user.id %}">Remove</a>
                         {% endif %}
                     </div>
                 </div>
